python/observability-online-eval/langgraph/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.43.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
Authlib==1.3.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dotenv-python==0.0.1
Flask==3.1.0
frozenlist==1.5.0
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langchain-weaviate==0.0.3
langgraph==0.2.62
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.25.1
maxim-py==3.1.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.59.7
orjson==3.10.14
packaging==24.2
pandas==2.2.3
propcache==0.2.1
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
simsimd==4.4.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
validators==0.34.0
weaviate==0.1.2
weaviate-client==4.10.4
Werkzeug==3.1.3
yarl==1.18.3
